difference-of-squares: fix inappropriate use of usize

+++ exercises/difference-of-squares/src/lib.rs
@@ -1,12 +1,12 @@
-pub fn square_of_sum(n: usize) -> usize {
+pub fn square_of_sum(n: u32) -> u32 {
     unimplemented!("square of sum of 1...{}", n)
 }
 
-pub fn sum_of_squares(n: usize) -> usize {
+pub fn sum_of_squares(n: u32) -> u32 {
     unimplemented!("sum of squares of 1...{}", n)
 }
 
-pub fn difference(n: usize) -> usize {
+pub fn difference(n: u32) -> u32 {
     unimplemented!(
         "difference between square of sum of 1...{n} and sum of squares of 1...{n}",
         n = n,
